+++ CHANGELOG.md
@@ -5,9 +5,8 @@
 - Added deadlock handling to batch commit method
 - Skip record from classify that have already been processed
 - Handle closed channel in rabbitMQ manager
-- Move Record status update to standard batch query
+- Move Record status update to standard batch query and prevent stale data
 - Handle `None` values in array of author names
-
 
 ## 2021-04-09 -- v0.5.4
 ### Fixed